requirements.txt
pycups
flask
flask-cors
reportlab
requests
waitress